+++ R/att_to_description.R
@@ -179,7 +179,7 @@ att_to_desc_from_pkg <- att_amend_desc
 #' imports = c("fcuk", "attachment"), suggests = c("knitr"))
 #' # In combination with other functions
 #' att_to_desc_from_is(path.d = file.path(dummypackage, "DESCRIPTION"),
-#' imports = att_from_rscripts(file.path(dummypackage, ".R")),
+#' imports = att_from_rscripts(file.path(dummypackage, "R")),
 #' suggests = att_from_rmds(file.path(dummypackage, "vignettes")))
 
 att_to_desc_from_is <- function(path.d = "DESCRIPTION", imports = NULL,